+++ app/src/main/java/bredesh/medico/Fragments/FragmentHome.java
@@ -118,7 +118,7 @@ public class FragmentHome extends Fragment {
             days[4] = c.getInt(c.getColumnIndex(db.THURSDAY));
             days[5] = c.getInt(c.getColumnIndex(db.FRIDAY));
             days[6] = c.getInt(c.getColumnIndex(db.SATURDAY));
-            float noOfRepetitions = c.getFloat(c.getColumnIndex(db.KEY_REPEATS));
+            int noOfRepetitions = c.getInt(c.getColumnIndex(db.KEY_REPEATS));
             String repetitionTpye = c.getString(c.getColumnIndex(db.KEY_REPETITION_TYPE));
 
             arrayList.add(index, new VideoItem(id, time, name, uri, days, noOfRepetitions, repetitionTpye, detailedTimes, allTimes, MedicoDB.KIND.Exercise));
